% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_layers.R
\name{base_game_boot_draw}
\alias{base_game_boot_draw}
\title{Draw the Base Layer for the Game Boot Plot}
\usage{
base_game_boot_draw(outline_switch = TRUE)
}
\arguments{
\item{outline_switch}{A conditional to display black outlines around the plot}
}
\value{

}
\description{
Draw the Base Layer for the Game Boot Plot
}
