% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_layers.R
\name{base_mode_select_draw}
\alias{base_mode_select_draw}
\title{Draw the Base layer for the Game Mode Plot}
\usage{
base_mode_select_draw(
  outline_switch = TRUE,
  inside_background,
  game_title = "GAMEPLAYER"
)
}
\arguments{
\item{outline_switch}{A conditional to display black outlines around the plot}

\item{inside_background}{The middle center background}

\item{game_title}{User Defined Title}
}
\value{

}
\description{
Draw the Base layer for the Game Mode Plot
}
