% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_layers.R
\name{base_title_draw}
\alias{base_title_draw}
\title{Draw the Base Layer for the Game Title Plot}
\usage{
base_title_draw(outline_switch = TRUE)
}
\arguments{
\item{outline_switch}{A conditional to display black outlines around the plot}
}
\value{

}
\description{
Draw the Base Layer for the Game Title Plot
}
