% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_layer_slider.R
\name{base_two_mode_battle_game_menu_1}
\alias{base_two_mode_battle_game_menu_1}
\title{Default Story Mode Text Location}
\usage{
base_two_mode_battle_game_menu_1(
  press_mode = TRUE,
  text = "STORY MODE",
  font = "Abduction2002",
  font_size = 8,
  scalar_x = 65,
  scalary = 80
)
}
\arguments{
\item{press_mode}{Toggle the mode switch}

\item{text}{User defined text for mode}

\item{font}{User defined font}

\item{font_size}{size of font}

\item{scalar_x}{numeric value to move menu text along x axis}

\item{scalar_y}{numeric value to move title photo along y axis}
}
\value{

}
\description{
Default Story Mode Text Location
}
