% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_battle_mode.R
\name{battle_mode_map_slots_function_4}
\alias{battle_mode_map_slots_function_4}
\title{Title}
\usage{
battle_mode_map_slots_function_4(
  number_1,
  number_2,
  number_3,
  number_4,
  name_1,
  name_2,
  name_3,
  name_4
)
}
\arguments{
\item{name_4}{}
}
\value{

}
\description{
Title
}
