% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initilize.R
\name{battle_story_game_init}
\alias{battle_story_game_init}
\title{Make the sub directory folder structure for a battle and story mode game}
\usage{
battle_story_game_init(path)
}
\arguments{
\item{path}{User specified working directory}
}
\value{

}
\description{
Make the sub directory folder structure for a battle and story mode game
}
