% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_select.R
\name{character_select_16}
\alias{character_select_16}
\title{Draw the 16 character battle menu}
\usage{
character_select_16(inside_background, media_ps)
}
\arguments{
\item{inside_background}{Background image for the select screen}

\item{media_ps}{List of pngs to load}
}
\value{

}
\description{
Draw the 16 character battle menu
}
