% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_select.R
\name{character_select_16_selected}
\alias{character_select_16_selected}
\title{Highlights 1 of 16 players in the 16 player select screen}
\usage{
character_select_16_selected(selected_number)
}
\arguments{
\item{selected_number}{Select a character slot number between 1 and 16}
}
\value{

}
\description{
Highlights 1 of 16 players in the 16 player select screen
}
