% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_select.R
\name{character_select_4}
\alias{character_select_4}
\title{Create the Story select screen for four characters}
\usage{
character_select_4(
  scalarx,
  media_p,
  media_ps,
  selected_number,
  first = "PLAYER",
  last = "NAME"
)
}
\arguments{
\item{scalarx}{scalar to change the amount of the profile picture is highlighted}

\item{media_p}{Player cutout to display}

\item{media_ps}{list of 4 profile pictures corresponding to the characters}

\item{selected_number}{number position between 1 and 4 to designate player to highlight}

\item{first}{First name label for player}

\item{last}{Last name label for player}
}
\value{

}
\description{
Create the Story select screen for four characters
}
