% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_vs.R
\name{character_vs_no_profiles}
\alias{character_vs_no_profiles}
\title{Create the character versus screen without profile photos}
\usage{
character_vs_no_profiles(
  player_1_cutout,
  player_2_cutout,
  insidebackground,
  label = "GAME PLAYER"
)
}
\arguments{
\item{player_1_cutout}{Path to cutout of player one character}

\item{player_2_cutout}{Path to cutout of player two character}

\item{insidebackground}{Path to background image for the vs screen}

\item{label}{Title photo for top and bottom}

\item{player_1_profile}{Path to profile picture for player one}

\item{player_2_profile}{Path to profile picture for player two}
}
\value{

}
\description{
Create the character versus screen without profile photos
}
