% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_game_setup.R
\name{create_game_device}
\alias{create_game_device}
\title{Creates game device for active game play}
\usage{
create_game_device()
}
\value{

}
\description{
Creates game device for active game play
}
