% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_screens.R
\name{game_boot}
\alias{game_boot}
\title{Boot up the Game and Bring in the Boot logo}
\usage{
game_boot(
  media_p,
  frames = 100,
  number,
  outline_switch = TRUE,
  output = F,
  path = NULL
)
}
\arguments{
\item{media_p}{Path to PNG file}

\item{frames}{number of frames to make the boot screen'}

\item{path}{}
}
\value{

}
\description{
Boot up the Game and Bring in the Boot logo
}
