% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_screens.R
\name{game_loading_screen_long_view}
\alias{game_loading_screen_long_view}
\title{Make a loading screen for the game}
\usage{
game_loading_screen_long_view(media_v, media_p, inside_background, nr)
}
\arguments{
\item{media_v}{Path to a video file for the loading screen}

\item{inside_background}{Background image for the game loading screen}

\item{nr}{Empty native raster object}

\item{media_ps}{list of paths to use for the logos}
}
\value{

}
\description{
Make a loading screen for the game
}
