% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{get_photo}
\alias{get_photo}
\title{Get the Photo}
\usage{
get_photo(my_path, name)
}
\arguments{
\item{my_path}{Directory to where the photos live}

\item{name}{Name of the photo}
}
\value{

}
\description{
Get the Photo
}
