% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{get_photo_random}
\alias{get_photo_random}
\title{Get a Random Photo}
\usage{
get_photo_random(my_path, my_pattern = "")
}
\arguments{
\item{my_path}{Directory to where the photos live}

\item{my_pattern}{optional argument to add regular expression search for file}
}
\value{

}
\description{
Get a Random Photo
}
