% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_sprites.R
\name{logo_left_bottom}
\alias{logo_left_bottom}
\title{Logo on the Bottom Left}
\usage{
logo_left_bottom(media_p, flop = FALSE, scalar_x = 0, scalar_y = 0)
}
\arguments{
\item{media_p}{Path to PNG file}

\item{flop}{Conditional switch in order to flop the image}

\item{scalar_x}{numeric value to move title photo along x axis}

\item{scalar_y}{numeric value to move title photo along y axis}
}
\value{

}
\description{
Logo on the Bottom Left
}
