% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_sprites.R
\name{logo_middle_center}
\alias{logo_middle_center}
\title{Logo in the Middle Center}
\usage{
logo_middle_center(media_p, flop = FALSE, scalar_x = 0, scalar_y = 0)
}
\arguments{
\item{media_p}{Path to PNG file}

\item{flop}{Conditional switch in order to flop the image}

\item{scalar_x}{numeric value to move title photo along x axis}

\item{scalar_y}{numeric value to move title photo along y axis}
}
\value{

}
\description{
Logo in the Middle Center
}
