% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_select.R
\name{map_select}
\alias{map_select}
\title{Select the map for the game}
\usage{
map_select(
  map_1,
  map_2,
  map_3,
  map_4,
  insidebackground,
  label_top = "GAME",
  label_bottom = "PLAYER"
)
}
\arguments{
\item{map_1}{Map 1}

\item{map_2}{Map 2}

\item{map_3}{Map 3}

\item{map_4}{Map 4}

\item{insidebackground}{Path to background image for the map screen}

\item{label_top}{Top text for the map}

\item{label_bottom}{Bottom text for the map}
}
\value{

}
\description{
Select the map for the game
}
