% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_select.R
\name{map_selected}
\alias{map_selected}
\title{Highlight the map the user selects}
\usage{
map_selected(
  map_1,
  map_2,
  map_3,
  map_4,
  insidebackground,
  label_top = "GAME",
  label_bottom = "PLAYER",
  selected_number
)
}
\arguments{
\item{map_1}{Map 1}

\item{map_2}{Map 2}

\item{map_3}{Map 3}

\item{map_4}{Map 4}

\item{insidebackground}{Path to background image for the map screen}

\item{label_top}{Top text for the map}

\item{label_bottom}{Bottom text for the map}

\item{selected_number}{Select a map slot number between 1 and 4}
}
\value{

}
\description{
Highlight the map the user selects
}
