% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_sprites.R
\name{menu_sprite_small_right}
\alias{menu_sprite_small_right}
\title{Draw a small sprite on the bottom right}
\usage{
menu_sprite_small_right(media_p, flop = FALSE, scalar_x = 0, scalar_y = 0)
}
\arguments{
\item{media_p}{Path to PNG file}

\item{flop}{Conditional switch in order to flop the image}

\item{scalar_x}{numeric value to move menu photo along x axis}

\item{scalar_y}{numeric value to move menu photo along y axis}
}
\value{

}
\description{
Draw a small sprite on the bottom right
}
