% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper_functions.R
\name{output_main_menu}
\alias{output_main_menu}
\title{Output the main menu screen to dir}
\usage{
output_main_menu(path, name, screen)
}
\arguments{
\item{path}{path to where the top level game working directory is}

\item{name}{name the user would like to give the screen with the default set as "title_screen"}

\item{screen}{variable with the stored plot}
}
\value{

}
\description{
Output the main menu screen to dir
}
