% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper_functions.R
\name{output_resized_sprites}
\alias{output_resized_sprites}
\title{Title}
\usage{
output_resized_sprites(path, name, number_1, number_2, image)
}
\arguments{
\item{number_2}{}
}
\value{

}
\description{
Title
}
