% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper_functions.R
\name{output_story_character_select}
\alias{output_story_character_select}
\title{Output the character select screen to dir}
\usage{
output_story_character_select(path, name = "Player", number = 1, screen)
}
\arguments{
\item{path}{path to where the top level game working directory is}

\item{name}{name the user would like to give the screen with the default set as "title_screen"}

\item{screen}{variable with the stored plot}
}
\value{

}
\description{
Output the character select screen to dir
}
