% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper_functions.R
\name{output_title}
\alias{output_title}
\title{Output the title screen to dir}
\usage{
output_title(path, name = "title_screen", screen)
}
\arguments{
\item{path}{path to where the top level game working directory is}

\item{name}{name the user would like to give the screen with the default set as "title_screen"}

\item{screen}{variable with the stored plot}
}
\value{

}
\description{
Output the title screen to dir
}
