% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper_functions.R
\name{output_vs}
\alias{output_vs}
\title{Output the character versus screen to dir}
\usage{
output_vs(
  path,
  name_1 = "Player",
  name_2 = "Player",
  number_1 = 1,
  number_2 = 1,
  screen
)
}
\arguments{
\item{path}{path to where the top level game working directory is}

\item{screen}{variable with the stored plot}

\item{name}{name the user would like to give the screen with the default set as "title_screen"}
}
\value{

}
\description{
Output the character versus screen to dir
}
