% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_layer_switches.R
\name{press_start_on}
\alias{press_start_on}
\title{Draw PRESS START on the Screen}
\usage{
press_start_on(press_start = TRUE, font = "Abduction2002")
}
\arguments{
\item{press_start}{Conditional to turn on the PRESS START switch}

\item{font}{Font used to draw "Press Start PRESS START}
}
\value{

}
\description{
Draw PRESS START on the Screen
}
