% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_select.R
\name{render_map}
\alias{render_map}
\title{Prepares map images for game play}
\usage{
render_map(path, map_number)
}
\arguments{
\item{path}{path to the map photo}

\item{map_number}{map image number}
}
\value{

}
\description{
Prepares map images for game play
}
