% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper_functions.R
\name{scale_image_height}
\alias{scale_image_height}
\title{Scale the image to the correct size for battle mode}
\usage{
scale_image_height(image_path)
}
\arguments{
\item{image_path}{path to the image}
}
\value{

}
\description{
Scale the image to the correct size for battle mode
}
