% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initilize.R
\name{video_game_init}
\alias{video_game_init}
\title{Creates video game development environment}
\usage{
video_game_init(path)
}
\arguments{
\item{path}{User specified working directory}
}
\value{

}
\description{
Creates video game development environment
}
