% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_screens.R
\name{game_boot}
\alias{game_boot}
\title{Boot up the Game and Bring in the Boot logo}
\usage{
game_boot(media_p, scalar_x = 0, scalar_y = 0)
}
\arguments{
\item{media_p}{Path to PNG file}

\item{scalar_x}{numeric value to move title photo along x axis}

\item{scalar_y}{numeric value to move title photo along y axis}
}
\value{

}
\description{
Boot up the Game and Bring in the Boot logo
}
